Move chunk tests to separate test file.
