+++ searchlib/CMakeLists.txt
@@ -114,6 +114,7 @@ vespa_define_module(
     src/tests/diskindex/fieldwriter
     src/tests/diskindex/fusion
     src/tests/diskindex/pagedict4
+    src/tests/docstore/chunk
     src/tests/docstore/document_store
     src/tests/docstore/document_store_visitor
     src/tests/docstore/lid_info
