+++ searchlib/src/tests/docstore/chunk/CMakeLists.txt
@@ -1,0 +1,8 @@
+# Copyright 2017 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_executable(searchlib_chunk_test_app TEST
+    SOURCES
+    chunk_test.cpp
+    DEPENDS
+    searchlib
+)
+vespa_add_test(NAME searchlib_chunk_test_app COMMAND searchlib_chunk_test_app)
