+++ searchlib/src/tests/docstore/chunk/FILES
@@ -1,0 +1,1 @@
+chunk_test.cpp
